+++ library/FieldUtils/ProcessModules/ProcessInterpPoints.cpp
@@ -38,7 +38,6 @@
 #include <boost/core/ignore_unused.hpp>
 #include <boost/geometry.hpp>
 #include <boost/lexical_cast.hpp>
-#include <boost/math/special_functions/fpclassify.hpp>
 
 #include <FieldUtils/Interpolator.h>
 #include <LibUtilities/BasicUtils/CsvIO.h>