+++ library/SolverUtils/Advection/AdvectionFR.cpp
@@ -44,7 +44,6 @@
 #include <MultiRegions/DisContField.h>
 #include <SolverUtils/Advection/AdvectionFR.h>
 #include <StdRegions/StdSegExp.h>
-#include <boost/math/special_functions/gamma.hpp>
 
 #include <iomanip>
 #include <iostream>
@@ -324,9 +323,9 @@ void AdvectionFR::SetupCFunctions(
                 NekDouble sign0 = pow(-1.0, p0);
 
                 // Factorial factor to build the scheme
-                NekDouble ap0 = boost::math::tgamma(2 * p0 + 1) /
-                                (pow(2.0, p0) * boost::math::tgamma(p0 + 1) *
-                                 boost::math::tgamma(p0 + 1));
+                NekDouble ap0 =
+                    std::tgamma(2 * p0 + 1) /
+                    (pow(2.0, p0) * std::tgamma(p0 + 1) * std::tgamma(p0 + 1));
 
                 // Scalar parameter which recovers the FR schemes
                 if (m_advType == "FRDG")
@@ -337,21 +336,20 @@ void AdvectionFR::SetupCFunctions(
                 {
                     c0 = 2.0 * p0 /
                          ((2.0 * p0 + 1.0) * (p0 + 1.0) *
-                          (ap0 * boost::math::tgamma(p0 + 1)) *
-                          (ap0 * boost::math::tgamma(p0 + 1)));
+                          (ap0 * std::tgamma(p0 + 1)) *
+                          (ap0 * std::tgamma(p0 + 1)));
                 }
                 else if (m_advType == "FRHU")
                 {
                     c0 = 2.0 * (p0 + 1.0) /
-                         ((2.0 * p0 + 1.0) * p0 *
-                          (ap0 * boost::math::tgamma(p0 + 1)) *
-                          (ap0 * boost::math::tgamma(p0 + 1)));
+                         ((2.0 * p0 + 1.0) * p0 * (ap0 * std::tgamma(p0 + 1)) *
+                          (ap0 * std::tgamma(p0 + 1)));
                 }
                 else if (m_advType == "FRcmin")
                 {
-                    c0 = -2.0 / ((2.0 * p0 + 1.0) *
-                                 (ap0 * boost::math::tgamma(p0 + 1)) *
-                                 (ap0 * boost::math::tgamma(p0 + 1)));
+                    c0 =
+                        -2.0 / ((2.0 * p0 + 1.0) * (ap0 * std::tgamma(p0 + 1)) *
+                                (ap0 * std::tgamma(p0 + 1)));
                 }
                 else if (m_advType == "FRcinf")
                 {
@@ -359,8 +357,8 @@ void AdvectionFR::SetupCFunctions(
                 }
 
                 NekDouble etap0 = 0.5 * c0 * (2.0 * p0 + 1.0) *
-                                  (ap0 * boost::math::tgamma(p0 + 1)) *
-                                  (ap0 * boost::math::tgamma(p0 + 1));
+                                  (ap0 * std::tgamma(p0 + 1)) *
+                                  (ap0 * std::tgamma(p0 + 1));
 
                 NekDouble overeta0 = 1.0 / (1.0 + etap0);
 
@@ -447,13 +445,13 @@ void AdvectionFR::SetupCFunctions(
                 NekDouble sign1 = pow(-1.0, p1);
 
                 // Factorial factor to build the scheme
-                NekDouble ap0 = boost::math::tgamma(2 * p0 + 1) /
-                                (pow(2.0, p0) * boost::math::tgamma(p0 + 1) *
-                                 boost::math::tgamma(p0 + 1));
+                NekDouble ap0 =
+                    std::tgamma(2 * p0 + 1) /
+                    (pow(2.0, p0) * std::tgamma(p0 + 1) * std::tgamma(p0 + 1));
 
-                NekDouble ap1 = boost::math::tgamma(2 * p1 + 1) /
-                                (pow(2.0, p1) * boost::math::tgamma(p1 + 1) *
-                                 boost::math::tgamma(p1 + 1));
+                NekDouble ap1 =
+                    std::tgamma(2 * p1 + 1) /
+                    (pow(2.0, p1) * std::tgamma(p1 + 1) * std::tgamma(p1 + 1));
 
                 // Scalar parameter which recovers the FR schemes
                 if (m_advType == "FRDG")
@@ -465,35 +463,33 @@ void AdvectionFR::SetupCFunctions(
                 {
                     c0 = 2.0 * p0 /
                          ((2.0 * p0 + 1.0) * (p0 + 1.0) *
-                          (ap0 * boost::math::tgamma(p0 + 1)) *
-                          (ap0 * boost::math::tgamma(p0 + 1)));
+                          (ap0 * std::tgamma(p0 + 1)) *
+                          (ap0 * std::tgamma(p0 + 1)));
 
                     c1 = 2.0 * p1 /
                          ((2.0 * p1 + 1.0) * (p1 + 1.0) *
-                          (ap1 * boost::math::tgamma(p1 + 1)) *
-                          (ap1 * boost::math::tgamma(p1 + 1)));
+                          (ap1 * std::tgamma(p1 + 1)) *
+                          (ap1 * std::tgamma(p1 + 1)));
                 }
                 else if (m_advType == "FRHU")
                 {
                     c0 = 2.0 * (p0 + 1.0) /
-                         ((2.0 * p0 + 1.0) * p0 *
-                          (ap0 * boost::math::tgamma(p0 + 1)) *
-                          (ap0 * boost::math::tgamma(p0 + 1)));
+                         ((2.0 * p0 + 1.0) * p0 * (ap0 * std::tgamma(p0 + 1)) *
+                          (ap0 * std::tgamma(p0 + 1)));
 
                     c1 = 2.0 * (p1 + 1.0) /
-                         ((2.0 * p1 + 1.0) * p1 *
-                          (ap1 * boost::math::tgamma(p1 + 1)) *
-                          (ap1 * boost::math::tgamma(p1 + 1)));
+                         ((2.0 * p1 + 1.0) * p1 * (ap1 * std::tgamma(p1 + 1)) *
+                          (ap1 * std::tgamma(p1 + 1)));
                 }
                 else if (m_advType == "FRcmin")
                 {
-                    c0 = -2.0 / ((2.0 * p0 + 1.0) *
-                                 (ap0 * boost::math::tgamma(p0 + 1)) *
-                                 (ap0 * boost::math::tgamma(p0 + 1)));
+                    c0 =
+                        -2.0 / ((2.0 * p0 + 1.0) * (ap0 * std::tgamma(p0 + 1)) *
+                                (ap0 * std::tgamma(p0 + 1)));
 
-                    c1 = -2.0 / ((2.0 * p1 + 1.0) *
-                                 (ap1 * boost::math::tgamma(p1 + 1)) *
-                                 (ap1 * boost::math::tgamma(p1 + 1)));
+                    c1 =
+                        -2.0 / ((2.0 * p1 + 1.0) * (ap1 * std::tgamma(p1 + 1)) *
+                                (ap1 * std::tgamma(p1 + 1)));
                 }
                 else if (m_advType == "FRcinf")
                 {
@@ -502,12 +498,12 @@ void AdvectionFR::SetupCFunctions(
                 }
 
                 NekDouble etap0 = 0.5 * c0 * (2.0 * p0 + 1.0) *
-                                  (ap0 * boost::math::tgamma(p0 + 1)) *
-                                  (ap0 * boost::math::tgamma(p0 + 1));
+                                  (ap0 * std::tgamma(p0 + 1)) *
+                                  (ap0 * std::tgamma(p0 + 1));
 
                 NekDouble etap1 = 0.5 * c1 * (2.0 * p1 + 1.0) *
-                                  (ap1 * boost::math::tgamma(p1 + 1)) *
-                                  (ap1 * boost::math::tgamma(p1 + 1));
+                                  (ap1 * std::tgamma(p1 + 1)) *
+                                  (ap1 * std::tgamma(p1 + 1));
 
                 NekDouble overeta0 = 1.0 / (1.0 + etap0);
                 NekDouble overeta1 = 1.0 / (1.0 + etap1);
@@ -629,19 +625,19 @@ void AdvectionFR::SetupCFunctions(
                 NekDouble sign1 = pow(-1.0, p1);
 
                 // Factorial factor to build the scheme
-                NekDouble ap0 = boost::math::tgamma(2 * p0 + 1) /
-                                (pow(2.0, p0) * boost::math::tgamma(p0 + 1) *
-                                 boost::math::tgamma(p0 + 1));
+                NekDouble ap0 =
+                    std::tgamma(2 * p0 + 1) /
+                    (pow(2.0, p0) * std::tgamma(p0 + 1) * std::tgamma(p0 + 1));
 
                 // Factorial factor to build the scheme
-                NekDouble ap1 = boost::math::tgamma(2 * p1 + 1) /
-                                (pow(2.0, p1) * boost::math::tgamma(p1 + 1) *
-                                 boost::math::tgamma(p1 + 1));
+                NekDouble ap1 =
+                    std::tgamma(2 * p1 + 1) /
+                    (pow(2.0, p1) * std::tgamma(p1 + 1) * std::tgamma(p1 + 1));
 
                 // Factorial factor to build the scheme
-                NekDouble ap2 = boost::math::tgamma(2 * p2 + 1) /
-                                (pow(2.0, p2) * boost::math::tgamma(p2 + 1) *
-                                 boost::math::tgamma(p2 + 1));
+                NekDouble ap2 =
+                    std::tgamma(2 * p2 + 1) /
+                    (pow(2.0, p2) * std::tgamma(p2 + 1) * std::tgamma(p2 + 1));
 
                 // Scalar parameter which recovers the FR schemes
                 if (m_advType == "FRDG")
@@ -654,49 +650,46 @@ void AdvectionFR::SetupCFunctions(
                 {
                     c0 = 2.0 * p0 /
                          ((2.0 * p0 + 1.0) * (p0 + 1.0) *
-                          (ap0 * boost::math::tgamma(p0 + 1)) *
-                          (ap0 * boost::math::tgamma(p0 + 1)));
+                          (ap0 * std::tgamma(p0 + 1)) *
+                          (ap0 * std::tgamma(p0 + 1)));
 
                     c1 = 2.0 * p1 /
                          ((2.0 * p1 + 1.0) * (p1 + 1.0) *
-                          (ap1 * boost::math::tgamma(p1 + 1)) *
-                          (ap1 * boost::math::tgamma(p1 + 1)));
+                          (ap1 * std::tgamma(p1 + 1)) *
+                          (ap1 * std::tgamma(p1 + 1)));
 
                     c2 = 2.0 * p2 /
                          ((2.0 * p2 + 1.0) * (p2 + 1.0) *
-                          (ap2 * boost::math::tgamma(p2 + 1)) *
-                          (ap2 * boost::math::tgamma(p2 + 1)));
+                          (ap2 * std::tgamma(p2 + 1)) *
+                          (ap2 * std::tgamma(p2 + 1)));
                 }
                 else if (m_advType == "FRHU")
                 {
                     c0 = 2.0 * (p0 + 1.0) /
-                         ((2.0 * p0 + 1.0) * p0 *
-                          (ap0 * boost::math::tgamma(p0 + 1)) *
-                          (ap0 * boost::math::tgamma(p0 + 1)));
+                         ((2.0 * p0 + 1.0) * p0 * (ap0 * std::tgamma(p0 + 1)) *
+                          (ap0 * std::tgamma(p0 + 1)));
 
                     c1 = 2.0 * (p1 + 1.0) /
-                         ((2.0 * p1 + 1.0) * p1 *
-                          (ap1 * boost::math::tgamma(p1 + 1)) *
-                          (ap1 * boost::math::tgamma(p1 + 1)));
+                         ((2.0 * p1 + 1.0) * p1 * (ap1 * std::tgamma(p1 + 1)) *
+                          (ap1 * std::tgamma(p1 + 1)));
 
                     c2 = 2.0 * (p2 + 1.0) /
-                         ((2.0 * p2 + 1.0) * p2 *
-                          (ap2 * boost::math::tgamma(p2 + 1)) *
-                          (ap2 * boost::math::tgamma(p2 + 1)));
+                         ((2.0 * p2 + 1.0) * p2 * (ap2 * std::tgamma(p2 + 1)) *
+                          (ap2 * std::tgamma(p2 + 1)));
                 }
                 else if (m_advType == "FRcmin")
                 {
-                    c0 = -2.0 / ((2.0 * p0 + 1.0) *
-                                 (ap0 * boost::math::tgamma(p0 + 1)) *
-                                 (ap0 * boost::math::tgamma(p0 + 1)));
+                    c0 =
+                        -2.0 / ((2.0 * p0 + 1.0) * (ap0 * std::tgamma(p0 + 1)) *
+                                (ap0 * std::tgamma(p0 + 1)));
 
-                    c1 = -2.0 / ((2.0 * p1 + 1.0) *
-                                 (ap1 * boost::math::tgamma(p1 + 1)) *
-                                 (ap1 * boost::math::tgamma(p1 + 1)));
+                    c1 =
+                        -2.0 / ((2.0 * p1 + 1.0) * (ap1 * std::tgamma(p1 + 1)) *
+                                (ap1 * std::tgamma(p1 + 1)));
 
-                    c2 = -2.0 / ((2.0 * p2 + 1.0) *
-                                 (ap2 * boost::math::tgamma(p2 + 1)) *
-                                 (ap2 * boost::math::tgamma(p2 + 1)));
+                    c2 =
+                        -2.0 / ((2.0 * p2 + 1.0) * (ap2 * std::tgamma(p2 + 1)) *
+                                (ap2 * std::tgamma(p2 + 1)));
                 }
                 else if (m_advType == "FRcinf")
                 {
@@ -706,16 +699,16 @@ void AdvectionFR::SetupCFunctions(
                 }
 
                 NekDouble etap0 = 0.5 * c0 * (2.0 * p0 + 1.0) *
-                                  (ap0 * boost::math::tgamma(p0 + 1)) *
-                                  (ap0 * boost::math::tgamma(p0 + 1));
+                                  (ap0 * std::tgamma(p0 + 1)) *
+                                  (ap0 * std::tgamma(p0 + 1));
 
                 NekDouble etap1 = 0.5 * c1 * (2.0 * p1 + 1.0) *
-                                  (ap1 * boost::math::tgamma(p1 + 1)) *
-                                  (ap1 * boost::math::tgamma(p1 + 1));
+                                  (ap1 * std::tgamma(p1 + 1)) *
+                                  (ap1 * std::tgamma(p1 + 1));
 
                 NekDouble etap2 = 0.5 * c2 * (2.0 * p2 + 1.0) *
-                                  (ap2 * boost::math::tgamma(p2 + 1)) *
-                                  (ap2 * boost::math::tgamma(p2 + 1));
+                                  (ap2 * std::tgamma(p2 + 1)) *
+                                  (ap2 * std::tgamma(p2 + 1));
 
                 NekDouble overeta0 = 1.0 / (1.0 + etap0);
                 NekDouble overeta1 = 1.0 / (1.0 + etap1);